requirements.txt

streamlit
Pillow
numpy
matplotlib
scikit-learn
opencv-python-headless
deepface
tensorflow
